Adds validation to input values and  background refresh.
